Restruktur lokasi file dan komponen
