+++ src/app/(authenticated user)/layout.tsx
@@ -1,17 +1,3 @@
-import type { ReactNode } from 'react'
-
-export default function Layout({
-  children,
-  navbar,
-}: Readonly<{
-  children: ReactNode
-  navbar: ReactNode
-}>) {
-  return (
-    <>
-      {navbar}
-
-      <main className="container mx-auto p-6">{children}</main>
-    </>
-  )
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return <div className="container mx-auto p-8">{children}</div>
 }
